+++ src/main/java/com/example/restfulapi/repositiories/PostRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostRepository extends PagingAndSortingRepository<Post, Integer> {
+public interface PostRepository extends PagingAndSortingRepository<Post, Long> {
 
     Page<Post> findAllByTitleContaining(String title, Pageable pageRequest);
 }
